Add line and fill widgets. Implement line hovering, clicking and selection, in list widget. Fix mouse movement event not being passed to hovered widgets. Fix glyph cropping and non-monospace meshing. Fix bitmapfontloader glyph sizes.

+++ source/voxelman/gui/events.d
@@ -58,7 +58,7 @@ struct PointerMoveEvent {
 	mixin GuiEvent!();
 }
 struct ScrollEvent {
-	ivec2 delta;
+	vec2 delta;
 	mixin GuiEvent!();
 }
